requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
diffusers==0.24.0
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gradio==3.35.2
gradio_client==0.8.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pipreqs==0.4.13
psutil==5.9.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pydyf==0.8.0
Pygments==2.17.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.15.2
safetensors==0.4.1
scipy==1.11.4
seaborn==0.13.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tinycss2==1.2.1
tokenizers==0.15.0
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.0.120
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.12
weasyprint==60.2
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
zopfli==0.2.3
